requirements.txt
openai>=1.0.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
schedule>=1.2.0
Pillow>=10.0.0
selenium>=4.15.0
webdriver-manager>=4.0.0
